+++ jvm-streaming/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+* Enabled StaticPointers extension to produce Static instances.
+
 ## [0.3] - 2018-02-27
 
 ### Added