+++ jvm-streaming/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+## [0.3.1] - 2018-01-01
+
+### Changed
+
 * Enabled StaticPointers extension to produce Static instances.
 * Fixed tests so they succeed when the build tree of jvm-batching is
   not available.